separate Partial View tests into new Fixture
